OPENEUROPA-1306-2: Refactoring the avportal module.
